Improve auto sign off if token is invalid
Navigator.pushAndRemoveUntil: This method clears the entire navigation stack (i.e., removes all previous screens) and pushes the new screen (in this case, the appropriate home screen or the login screen).
(Route<dynamic> route) => false: This argument ensures that all previous routes are removed from the stack.
Session Validation: The session is checked, and depending on the token and role, the user is navigated to the appropriate page. If the token is invalid or expired, the user is redirected to the sign-in page, and all previous pages are removed from the stack.

After successful token validation, the user is navigated to their respective home screen (Admin, Manager, or Driver) while clearing all previous screens in the stack.
If the session is expired or invalid, the user is taken to the SignInPage with the entire navigation stack cleared.
This ensures that the user is taken to the main screen (login or splash) after an automatic sign-out, effectively "deleting" all the previous widget trees.